+++ src/modules/components/page-renderer/component-types/icon/style.tsx
@@ -1,5 +1,6 @@
 import styled from "@emotion/styled";
 import { MEDIUM_GREY } from "../../../../../styles/colors";
+import { ACESSIBILITY_RED } from "./../../../../../styles/colors";
 
 export const IconOverlaySpan = styled(
   ({ isRendererPage, children, ...rest }: any) =>
@@ -12,9 +13,40 @@ export const IconOverlaySpan = styled(
   color: ${MEDIUM_GREY};
   font-size: 20px;
   margin: 8px;
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 
-  &:hover svg {
-    color: grey;
+  & #delete-icon-bg {
+    position: absolute;
+    display: none;
+    justify-content: center;
+    align-items: center;
+    background-color: ${ACESSIBILITY_RED};
+    opacity: 1;
+    width: 20px;
+    height: 20px;
+    border-radius: 50%;
+    top: -2px;
+    right: -2px;
+    z-index: 1;
+    cursor: pointer;
+  }
+
+  & #delete-icon {
+    color: white;
+    opacity: 1;
+    font-size: 16px;
+  }
+
+  &:hover #app-icon path {
+    color: ${ACESSIBILITY_RED};
     opacity: 0.5;
+  }
+
+  &:hover #delete-icon-bg {
+    display: flex;
+    opacity: 1;
   }
 `;